requirements.txt
dj-database-url==0.5.0
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-spectacular==0.17.3
gunicorn==20.1.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
six==1.16.0